Bump R version and switch base Docker to versioned R
It seems that some changes in Debian testing, the base OS for the r-base
Docker images, violated some expectations for pip and Cython. svtk no
longer installs as pip reports that it cannot find Cython. Therefore, we
switch the base image to a versioned R from the rocker project which is
built on Ubuntu LTS.